Use `babel`'s built-in option manager for loading configs.
There is an issue if you configure this plugin anywhere outside of your `.babelrc`. For example, if you have:

```json
{
  "presets": ["my-preset"]
}
```

And in your preset you have this plugin configured, then this plugin will fail to find that configuration value. By leveraging babel's built-in option manager for doing all the heavy lifting we avoid these pitfalls.